Add files via upload
Problem Statement: Predict the percentage of an student based on the no. of study hours.

Working: The data is first visually represented and then a train test split is preformed further which a simple linear regression is fit to the train data and then used for further prediction of  scores for the test data. The model achieved a R-squared value of 0.97 and Root Mean Squared error of 4.35

The model also predicted score 92.39% for an user input of 9.25 hours/day study time.